trehacklab/requirements.txt
Mezzanine==5.1.0
fontawesome-free==5.15.4
django-allauth==0.47.0